refactor: move x86 I/O port specs to vstd_extra
