+++ verified_libs/vstd_extra/src/external/mod.rs
@@ -7,6 +7,8 @@ pub mod convert;
 pub mod deref;
 pub mod ilog2;
 pub mod int_specs;
+#[cfg(target_arch = "x86_64")]
+pub mod io_port;
 pub mod nonnull;
 pub mod ptr;
 pub mod range;
@@ -17,6 +19,8 @@ pub mod time;
 pub use btree::*;
 pub use ilog2::*;
 pub use int_specs::*;
+#[cfg(target_arch = "x86_64")]
+pub use io_port::*;
 pub use nonnull::*;
 pub use ptr::*;
 pub use range::*;
